improve graphql error handling

+++ api/app/graphql/mutations/delete_issue_comment.rb
@@ -1,20 +1,19 @@
 # frozen_string_literal: true
 
 module Mutations
-  class DeleteIssueComment < GraphQL::Schema::RelayClassicMutation
-    field :errors, [String], null: false
-
+  class DeleteIssueComment < BaseMutation
     argument :issue_coment_id, ID, required: true
 
     def resolve(issue_coment_id:)
-      issue_comment = IssueComment.find_by!(id: issue_coment_id)
+      issue_comment = IssueComment.find_by(id: issue_coment_id)
+      raise RecordNotExists.new(IssueComment, id: issue_coment_id) if issue_comment.nil?
+
       Acl.permit!(mutation: self, args: { issue_comment: issue_comment })
 
       if issue_comment.destroy
-        # errorsが空なら成功とする
-        { errors: [] }
+        {}
       else
-        { errors: issue_comment.errors.full_messages }
+        add_errors(issue_comment)
       end
     end
   end
