+++ api/app/graphql/mutations/delete_notice.rb
@@ -1,20 +1,19 @@
 # frozen_string_literal: true
 
 module Mutations
-  class DeleteNotice < GraphQL::Schema::RelayClassicMutation
-    field :errors, [String], null: false
-
+  class DeleteNotice < BaseMutation
     argument :notice_id, ID, required: true
 
     def resolve(notice_id:)
-      notice = Notice.find_by!(id: notice_id)
+      notice = Notice.find_by(id: notice_id)
+      raise RecordNotExists.new(Notice, id: notice_id) if notice.nil?
+
       Acl.permit!(mutation: self, args: { notice: notice })
 
       if notice.destroy
-        # errorsが空なら成功とする
-        { errors: [] }
+        {}
       else
-        { errors: notice.errors.full_messages }
+        add_errors(notice)
       end
     end
   end
